// front-end/src/api/api.ts
export interface Event {
    /**
     * Name of the person that created an event.
     * @type {string}
     * @memberof Event
     */
    'firstName': string;
    /**
     * Last name of the person that created an event.
     * @type {string}
     * @memberof Event
     */
    'lastName': string;
    /**
     * Email the person that created an event.
     * @type {string}
     * @memberof Event
     */
    'email': string;
    /**
     * Date when an event occured.
     * @type {string}
     * @memberof Event
     */
    'eventDate': string;
}